Add scheme short and long names to output file attrs
Also fix units in output irradiance and actinic flux vars.
Currently they are output as W/m^2 in band. Dividing by the band
widths would give the distribution variable, which is what we would use
to plot the spectrum, so perhaps should change what is output, but
for now just changing the units for accuracy.